changed parameter scaling matrix to 1.0 for all cases - seems to be too strong for 10par_xsec problem when one parameter has vastly different variance than others.

+++ pyemu/en.py
@@ -27,7 +27,7 @@ class Ensemble(pd.DataFrame):
         self.__mean_values = mean_values
 
     def as_pyemu_matrix(self):
-        x = self.copy().as_matrix()
+        x = self.copy().as_matrix().astype(np.float)
         return Matrix(x=x,row_names=list(self.index),
                       col_names=list(self.columns))
